+++ tests/data/post_processing_test.py
@@ -1,0 +1,2 @@
+def post_processing(finder):
+    return {"quantity": 5, "another_quantity": 10}